backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
motor>=3.3.0
pymongo>=4.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
google-generativeai>=0.8.0
python-multipart>=0.0.6
Pillow>=10.0.0
requests>=2.31.0
PyYAML>=6.0
